requirements.txt
Django==4.1.3
djangorestframework==3.15.1
drf_social_oauth2==2.2.0
python-dotenv==1.0.1
django-cors-headers==4.3.1